Backend/requirements.txt
Django>=5.2.0
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
python-dotenv>=1.0.0
djangorestframework
django-cors-headers
requests>=2.31.0
djangorestframework-simplejwt
django-otp>=1.1.3
django-two-factor-auth>=1.15.2
pyotp>=2.8.0
qrcode>=7.4.2
Pillow>=10.0.0
elasticsearch==7.13.4
web3